+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * get multiple metadata entry value helper methods
 * `InMemoryGraph` bug fixes
 * CLI improvements
+* in process Elasticsearch
 
 # v2.0.3